refactor: Clean Architecture 변경
