+++ lib/clean/clean_main.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
-
-import 'package:image_search_clean/clean/data/pixabay_api.dart';
-import 'package:image_search_clean/clean/ui/home_screen.dart';
-import 'package:image_search_clean/clean/ui/home_view_model.dart';
 import 'package:provider/provider.dart';
+import 'package:http/http.dart' as http;
+import 'package:image_search_clean/clean/data/data_source/pixabay_api.dart';
+import 'package:image_search_clean/clean/data/repository/pixabay_api_repository_impl.dart';
+import 'package:image_search_clean/clean/presentation/home/home_screen.dart';
+import 'package:image_search_clean/clean/presentation/home/home_view_model.dart';
 
 void main() {
   runApp(const MyApp());
@@ -22,7 +23,7 @@ class MyApp extends StatelessWidget {
         appBarTheme: const AppBarTheme(centerTitle: true),
       ),
       home: ChangeNotifierProvider(
-        create: (_) => HomeViewModel(PixabayApi()),
+        create: (_) => HomeViewModel(PixabayApiRepositoryImpl(PixabayApi(http.Client()))),
         child: const HomeScreen(),
       ),
     );
