+++ lib/clean/data/data_source/pixabay_api.dart
@@ -1,0 +1,20 @@
+import 'dart:convert';
+
+import 'package:http/http.dart' as http;
+
+class PixabayApi {
+  final http.Client client;
+
+  PixabayApi(this.client);
+
+  static const baseUrl = 'https://pixabay.com/api/';
+  static const key = '10711147-dc41758b93b263957026bdadb';
+
+  Future<Iterable> fetch(String query) async {
+    final response = await client
+        .get(Uri.parse('$baseUrl?key=$key&q=$query&image_type=photo'));
+    Map<String, dynamic> jsonResponse = jsonDecode(response.body);
+    Iterable hits = jsonResponse['hits'];
+    return hits;
+  }
+}
