+++ lib/clean/domain/model/photo.dart
