+++ lib/clean/domain/repository/photo_api_repository.dart
@@ -1,4 +1,4 @@
-import 'package:image_search_clean/clean/model/photo.dart';
+import 'package:image_search_clean/clean/domain/model/photo.dart';
 
 abstract class PhotoApiRepository {
   Future<List<Photo>> fetch(String query);
