+++ lib/clean/presentation/home/components/photo_widget.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-
-import 'package:image_search_clean/clean/model/photo.dart';
+import 'package:image_search_clean/clean/domain/model/photo.dart';
 
 class PhotoWidget extends StatelessWidget {
   final Photo photo;
