+++ test/clean/ui/home_view_model_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:image_search_clean/clean/domain/model/photo.dart';
+import 'package:image_search_clean/clean/domain/repository/photo_api_repository.dart';
+import 'package:image_search_clean/clean/presentation/home/home_view_model.dart';
 
-import 'package:image_search_clean/clean/data/photo_api_repository.dart';
-import 'package:image_search_clean/clean/model/photo.dart';
-import 'package:image_search_clean/clean/ui/home_view_model.dart';
 
 void main() {
   test('Stream Test', () async {
